SVG element selection refactoring
